refactor : delete package seller under dto
