+++ seller-api/src/main/java/zerobase/sellerapi/dto/SellerEditDto.java
@@ -1,10 +1,13 @@
-package zerobase.sellerapi.dto.seller;
+package zerobase.sellerapi.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+/**
+ * 판매자 정보 수정을 위한 Dto
+ */
 @Getter
 @Builder
 @NoArgsConstructor
