+++ seller-api/src/main/java/zerobase/sellerapi/dto/SellerSignUpDto.java
@@ -1,4 +1,4 @@
-package zerobase.sellerapi.dto.seller;
+package zerobase.sellerapi.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
